XAI-CyberSec-main/NewCadetsE3/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.10=pypi_0
aiosignal=1.3.1=pypi_0
attrs=24.2.0=pypi_0
bzip2=1.0.8=h99b78c6_7
ca-certificates=2024.8.30=hf0a4a13_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
cpython=3.13.0=py313hd8ed1ab_100
cyrus-sasl=2.1.27=h60b93bd_7
expat=2.6.3=hf9b8971_0
filelock=3.16.1=pyhd8ed1ab_0
frozenlist=1.5.0=pypi_0
fsspec=2024.10.0=pyhff2d567_0
gmp=6.3.0=h7bae524_2
gmpy2=2.1.5=py313he559d9b_2
icu=75.1=hfee45f7_0
idna=3.10=pypi_0
jinja2=3.1.4=pyhd8ed1ab_0
joblib=1.4.2=pypi_0
krb5=1.21.3=h237132a_0
libabseil=20240722.0=cxx17_hf9b8971_1
libblas=3.9.0=25_osxarm64_openblas
libcblas=3.9.0=25_osxarm64_openblas
libcxx=19.1.2=ha82da77_0
libedit=3.1.20191231=hc8eb9b7_2
libexpat=2.6.3=hf9b8971_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
liblapack=3.9.0=25_osxarm64_openblas
libmpdec=4.0.0=h99b78c6_0
libntlm=1.4=h3422bc3_1002
libopenblas=0.3.28=openmp_h517c56d_0
libpq=17.0=h9fd3c6c_4
libprotobuf=5.28.2=h8f0b736_0
libsqlite=3.47.0=hbaaea75_0
libtorch=2.4.1=cpu_generic_h6570967_3
libuv=1.49.2=h7ab814d_0
libzlib=1.3.1=h8359307_2
llvm-openmp=19.1.2=hb52a8e5_0
markupsafe=3.0.2=py313heb2b014_0
mpc=1.3.1=h8f1351a_1
mpfr=4.2.1=hb693164_3
mpmath=1.3.0=pyhd8ed1ab_0
multidict=6.1.0=pypi_0
ncurses=6.5=h7bae524_1
networkx=3.4.2=pyhd8ed1ab_0
nomkl=1.0=h5ca1d4c_0
numpy=2.1.2=py313hab0c69d_0
openldap=2.6.8=h50f2afc_0
openssl=3.3.2=h8359307_0
pip=24.2=pyh145f28c_1
propcache=0.2.0=pypi_0
psutil=6.1.0=pypi_0
psycopg2=2.9.10=pypi_0
pyparsing=3.2.0=pypi_0
python=3.13.0=h75c3a9f_100_cp313
python-graphviz=0.20.3=pypi_0
python_abi=3.13=5_cp313
pytorch=2.4.1=cpu_generic_py313h37cab96_3
pytz=2024.2=pypi_0
readline=8.2=h92ec313_1
requests=2.32.3=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=pyhd8ed1ab_0
sleef=3.7=h7783ee8_0
sqlite=3.47.0=hcd14bea_0
sympy=1.13.3=pyh2585a3b_104
threadpoolctl=3.5.0=pypi_0
tk=8.6.13=h5083fa2_1
torch-cluster=1.6.3=pypi_0
torch-geometric=2.6.1=pypi_0
torch-scatter=2.1.2=pypi_0
torch-sparse=0.6.18=pypi_0
torch-spline-conv=1.2.2=pypi_0
tqdm=4.66.5=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024b=hc8b5060_0
urllib3=2.2.3=pypi_0
wheel=0.44.0=pyhd8ed1ab_0
xxhash=3.5.0=pypi_0
xz=5.2.6=h57fd34a_0
yarl=1.16.0=pypi_0
zlib=1.3.1=h8359307_2
